requirements.txt
streamlit
mysql-connector-python
python-dotenv

pandas
plotly
